requirements.txt
requests==2.27.1
beautifulsoup4===4.10.0
Pillow===9.0.1
python-dotenv===0.19.2
requests-html===0.10.0
spotipy==2.19.0